First attempt at a jump action.
Need to refine the action so it's not just the first segment leading the
jump, because the other segments don't have their own force applied and end up
in a rotation motion as they try to keep up.

Poor worm ends up doing somersaults in the air and getting all tangled up.

Also, refined the code so segments keep up with where the forward segment
*will* not, not where it currently is, because it may start spazzing out
because it's never able to "catch up".