Some fixes to make segments less glitchy.
Temporarily "disattach" segment from surface if a large instantaneous velocity
correction is being applied to "keep up" with the segment in front.

Also, some minor updates for when behaviour is triggered (sometimes only do
something if on a surface, or not on a surface).

Also, more strict avoidance of jack-knifing.  Worm was contorting a lot.