Added precise targeting for bugs.
You have no idea how long it took me to figure out how to make this work properly.

One thing I learned from this:  There's a difference in behaviour depending
on the order of when you add a node as a child, and when you set the node's
position and velocity.

In my case, I needed to add the slime as child node first, *then* set the
global position and linear velocity.  Otherwise, the coordinates get all wonky.