More work on options menu.
Need to dynamically create it and add it as a sibling for other menus, so the
show / hide logic doesn't conflict between them?

I don't feel like this is an ideal way to handle sub-menus, but it's good
enough for now.

Maybe once I have more experience I can go back and implement a cleaner
solution.