Add an option to change the worm's colour scheme.
You would *think* that would be a simple update to make - replace
one set of colours with another at runtime.

However, it turned out to be a rather complicated update.

A breakdown of what was involved:

From what I understand, the best way to dynamically change the
colour of a sprite is to modify its "modulate" property.  For
effective colour updates, the starting colour of the sprite should
be white & transparent with the appropriate shape.  Each differently
coloured part would have to be stored in a separate sprite so they
could be modulated independently, then combined as siblings in the
scene tree.  I had to re-work all the worm sprites to be separated
in this way.

Unfortunately, this further complicates the walking animation.  The
legs are multiple colours (including the outline) that are now
split into multiple images to be colour modulated, but
AnimatedSprite2D objects can only use individual image files as
frames, not nested sprite objects.  So I had to switch to using the
more generic AnimationPlayer to explicitly switch between the
groups of sprites forming the frames of the animation.

Not to mention the extra global state needed to hold the colours,
and managing the signal to send to the worm parts to tell them
when the colours have been updated.

Fortunately, Godot includes a handy ColorPickerButton object for
getting a colour selection from the user, so at least I didn't
have to hand-roll *that* part of the interface.