Turn off auto-targeting of bugs.
User must figure out where to aim for themselves.

While it was interesting to implement the solver to intersect the parabolic
curve of the slime with the oscillating paths of the bugs, using the auto-aim
made the game... boring.  There's already very little in terms of gameplay,
and auto-aim was robbing the player of any engagement.

I left a switch to turn it on from the code, just in case there's some need
for it in the future (accessibility maybe?)

Also, added WASD as alternative to arrow keys for movement.