While holding down a movement key, keep moving even if the worm's direction is changing.
E.g., if turning a corner to go up a wall, or even upside down, just keep going
so long as the original movement key is held.

This removes some of the awkwardness in having to switch keys while moving
quickly around different surface orientations.

A playtester recommended binding the direction keys to the worm's body
direction to make it more "sonic-esque", e.g. if moving up a wall the
"D" / "right" key would move it up the wall, and "A" / "left" would move it
down the wall.  I'm still not sure if that's the right approach, since walking
on the ceiling would have counter-intuitive controls ("left" moves right and
"right" moves left).

This current approach is hopefully a good compromise, where as long as you're
holding down the "D" / "right" key it will move in a sonic-esque way when
reaching a wall.  If you  let go of the key, the you'd have to use the
appropriate direction to get moving again.