Some visual fixes for peer worms.
Make sure z index is high enough to be drawn in front of background.
Make sure worm stays oriented (pieces not rolling with the collision circles).
Make sure colours get updated properly (and independent of local worm).

Note: Still a lot of things to fix.  Peer worms twitch on screen, and the
eating area still seems to be active???