Use MultiplayerSynchronizer on the sprite TileMapLayer, so the bugs should now be synchronized between client and server.
Note: For this to work properly, it seems that the names of the bug scenes
created from these tiles need to be fixed up.  The default names aren't working
very well in the synchronizer code.

Also note: this doesn't deal with user interactions with the bugs (eating and
sliming).

Also also note: there's a bit of a visual glitch happening on the client side,
where the position of the bugs flips rapidly between two states?  Pausing and
unpausing the game seems to fix that.  Need to investigate what exactly is
happening.