Bugs can now get eaten by a worm in multiplayer, and it actually gets "eaten", i.e. disappears from the scene.
Disconnecting and re-connecting the client confirms that the bug is still gone.

Note: sliming the bug isn't working in multiplayer, this is just literally
jumping up and eating it in mid-air.

Also, it doesn't update the score yet, because the multiplayer worm spawned
in the "Worms" node hasn't had its signals connected to anything yet.

Might remove the score display for multiplayer anyway, and just let the bugs
respawn periodically to be re-eaten by another player just for something to do.