Slime shooting now working in multiplayer.
Had to refactor things so the slime gets spawned by the Screen scene instead
of by the WormFront.  This is because the WormFront is controlled by the user,
but the slimes need to be managed by the server to deal with collision logic.

This requires sending a new signal from the worm to request slime be created.