Allow players to die and respawn in multiplayer.
The trickiest part was in WormSegment.respawn().  For the longest time, I was
struggling with trying to set the global_position back to the origin.  It
wasn't "seeing" the updated values and was leaving the worm back in the area
where it died.  Turns out I have to use "set_deferred" because I'm setting
the global_position outside the segments's own _process call, and the value
just gets overwritten by the next process call... I think?

Anyway, it works now, so committing before something else breaks!