Remove a "quick fix" that was applied when first joining a server.
Now I understand that the bugs were glitching on the client side when first
joining because the game was trying to instantaneously update the bug positions
and velocities on the client to match the server, which was tripping up the
physics processing.

The quick fix of briefly pausing the game worked because it stopped the physics
during the initial synchronization of properties.  By the time of unpausing,
the client / server versions of the bug positions and velocities were identical
(or close enough) so the physics wouldn't see any more unexpected changes to
the values for the rest of the game.

The synchronization fix from the previous commit seems to have also fixed this
initial synchronization problem.

It feels good to finally understand this issue, but holy hell that was a
mentally painful learning experience to work through.

Anyway... learning!