Test multiplayer connection before starting Screen scene.
That way, if server is unavailable (or invite code in invalid), no game
is started.

Also allows a proper error message to be printed when the connection
fails.

Note: doing the connection test is causing some error messages in the
debug console, because the server is immediately trying to sync
information to the client even though the client hasn't loaded the Screen
scene yet.

For now, just going to accept those error messages.  Could revisit the
approach if I learn more useful information about how to work with
connections.