Fix a weird problem with using is_multiplayer_authority() as a condition for synchronization.
In a single player game, this works ok - we are our own authority.
In a multiplayer game, this works ok - client is not authority, it will
listen for incoming property updates.
However, if a multiplayer game is started, then quit, then a single
player game is started, this condition no longer works (?)
The single player session no longer indicates it's the authority, so
the bug locations no longer get updated properly.

Use a slightly different test condition, which seems to be more robust
when switching between local / multiplayer games.