Testing out a different font for the chat interface (RoundedMplus1c-Medium).
This one is rounded like the Proxima Soft Semi-bold used in Goodbye Volcano
High, but is free to use.  Not a total match, but hopefully close enough?

Just a note if anyone is cloning this repo: this branch may be rebased if I
end up choosing a different font.  I'm not going to keep a history of font
choices if I keep changing my mind.  These ttf files take up a relatively large
amount of space - doubling the size of the project with just the one font!

*Update* April 10, 2025:
I did, in fact, rebase this branch and remove the font file.  It's too big.
I put the download location in the file fonts/roundedmplus1c/SOURCE.
I will keep the font file (untracked) in my local working directory, so it
will still be included when I do builds.