Set width of chat window to the maximum needed.
Before, it would shrink to the size of the current messages. So if the chat
started with some short messages, then a larger one arrived, there was a
visible shift in horizontal alignment of all the messages on screen.

A minor issue but annoying.

Also, for posterity, found out the hard way that a Container's "size" property
is not available within the _ready function.

Also, querying the sizes doesn't work when the element is hidden (not displayed
on screen), so have to wait until after _ready, and until after the chat
becomes visible.