Dynamically set chat font at runtime.
Font files aren't included in the repository, but can be added manually.

Note: I *did* try using a Theme to contain these changes, but there was an
issue with using a Theme where the font is chosen dynmically (based on
available font file).

I *could* set a "font" and "default_font" for the theme to fall back when the
preferred font isn't there, BUT it spews a whole bunch of error messages in
the process.

I tried setting the "font" attribute of the Theme from the Globals _ready
function, but the "WORM CHAT" label didn't see the updates and continued to
use the Godot default font (guess it renders before Globals._ready is called?)