When opening menus / chat in multiplayer, keep screen running in the background.
Can't do a true "pause" in that context, and weird things happen if you try to
pause.  For instance, if the player's worm was on a moving platform and then
when to the chat menu, then their worm would be frozen in place and not
continue to move with the platform.  Upon returning to the game, the worm
would then spaz out.

Now things continue to run and update as they should, while ignoring movement
keys while the user is navigating the menu systems.