Updated server code.
Servers are now started on the command-line with the --headless option.

Disabled the "Create Server" button from the GUI, since it's not needed in most
cases.

Also, removed my handle and the localhost invite code from the default contents
of the multiplayer menu.

Also also, create and display the invite code when starting a server.