Need to limit the max fps for headless servers.
Without a limit imposed, and without any display to render, the fps shoots way
up.  This causes the flying bugs to move in a very glitchy way for some reason.

This problem took a long time to identify.  For a while, I thought there was
an issue with something in the scene not being initialized properly in headless
mode, or perhaps an incompatibility between headless mode and
WebSocketMultiplayerPeer or MultiplayerSynchronizer.

Note: This rate limit is only needed for headless server builds, but I'm
putting it permanently in the project settings or I'll keep forgetting to do
it for the server build.