Update the status of *all* karts for a multiplayer peer when they first join, not just the taken karts.
There was a bug where it was only updating the status of karts that were
taken for this current race, but not clearing the status of untaken karts.
So if a kart was currently untaken, but was taken in another multiplayer
race that was previous played, then it would still show up as taken.