Added nail polish and beetles to multiplayer games.
Also, I may have figured out why I was having trouble using the
RaceSynchronizer nodes with item blocks... if I "instantiate" that node in a
scene such as nail polish or beetle, it doesn't save the properties I'm trying
to sync.  However, if I just add it as a child node, it works.

So maybe I was "instantiating" the RaceSynchronizer on the D20 object, which
was causing the lack of updates?

Oh well... at least I got item blocks working with RPC calls anyway.