Got skid marks to work on multiplayer games.
Not very efficient, in the sense that it needs to continuously synchronize
the whole set of points over the network.  But it works, and requires minimal
code changes.

Doing something more bandwidth-friendly would require a LOT of refactoring for
the skid mark code, and I really don't want to get into that this late in the
development cycle.