For multiplayer game menu, don't fill text box with the default name "Player".
Put that as placeholder text instead, and leave actual text box blank to start.

This should make it easier for mobile users, so they don't have to awkwardly
tap the on-screen backspace key a bunch of times to remove the "Player" text.
This is extra awkward because Godot's experimental on-screen keyboard covers
up the text box, so it isn't clear when the box was actually cleared.

Now that it starts blank, check if it's still blank when user tries to start
or join a multiplayer race.