Fix a glitch in multiplayer where sometimes the cars are displayed in the wrong positions during the initial countdown.
The workaround is to always sync the positions, even if cars aren't moving.

I guess sometimes the client joins just after the server has arranged the cars,
and wasn't getting the initial updated positions (which were sent only once
as a delta).

Car positions are changing 99% of the time anyway, so using sync "on change"
doesn't make sense anyway.