Fix a glitch in snowy track, where the falling snow doesn't appear right away.
The problem was a combination of how the track coords are offset to some far
distance (to avoid spurrious collisions with the main menu elements), and how
the snow texture was wrapped on the screen when the camera moved.

The initial move of the camera was too large for a single screen wrap, so it
was taking multiple frames for it to get corrected.  Now it forces the wrapping
to work within the initial frame using a "while" loop.

I could probably just use modular arithmetic, now that I think about it...