Fix a rare bug where if the player happens to cross the finish line while the "are you lost?" arrow is active, then the arrow stays visible.
It stays at a fixed angle w.r.t. the car (rotates as the car rotates), and is
kind of strange to have there.  I don't even know why the arrow would be active
when crossing the finish line (you're obviously close enough to the track to
make progress?).  But it's easy enough to fix anyway.